Sat, Oct 21, 2023, 6:25 AM +00:00

+++ readability.c
@@ -25,7 +25,7 @@ int main(void)
     {
         sentences++;
     }
-    float L = letters / words * 100
-    float S = sentences / words * 100
+    float L = (float)letters / (float)words * 100
+    float S = (float)sentences / (float)words * 100
     int index = 0,0588 * L - 0,296 * S - 15,8;
 }
